[dwc] Added save retention value for lpddr4.

+++ dwc/include/dwc_dram_param.h
@@ -2814,7 +2814,7 @@ void dwc_ddrphy_phyinit_userCustom_G_waitFwDone ();
 void run_customPostTrain();
 int ctl_trigger_init_and_wait_normal();
 void dwc_ddrphy_apb_wr(UINT32 adr, UINT32 dat);
-int dwc_ddrphy_apb_rd(UINT32 adr);
+UINT16 dwc_ddrphy_apb_rd(UINT32 adr);
 int ReadSDSector(
 	unsigned int sectorNo,
 	unsigned int pageCount,
